modify ivim

+++ setup.sh
@@ -35,6 +35,10 @@ function set_shell()
   fi
 }
 
+#main
+#start config vim and alias
+#
+
 if which apt-get >/dev/null; then
   sudo apt-get install -y ctags ack cscope
 elif which yum >/dev/null; then
@@ -46,17 +50,20 @@ fi
 
 sudo ln -s /usr/bin/ctags /usr/local/bin/ctags
 
-#backup old_vim
-if [ -d ~/.vim ]; then
-  mv -f ~/.vim ~/vim_old
-fi
+#exist ivim do't backup and git clone
+if [ ! -f ~/.vim/setup.sh ];then
+  #backup old_vim
+  if [ -d ~/.vim ]; then
+    mv -f ~/.vim ~/vim_old
+  fi
 
-if [ -f ~/.vimrc ];then
-  mv -f ~/.vimrc ~/.vimrc_old
-fi
+  if [ -f ~/.vimrc ];then
+    mv -f ~/.vimrc ~/.vimrc_old
+  fi
 
-#ivim config
-cd ~/ && git clone https://github.com/lostedleo/ivim.git ~/.vim
+  #ivim config
+  cd ~/ && git clone https://github.com/lostedleo/ivim.git ~/.vim
+fi
 
 #set shell config
 set_shell
@@ -65,13 +72,17 @@ set_shell
 ln -sf ~/.vim/vimrc ~/.vimrc
 ln -sf ~/.vim/screenrc ~/.screenrc
 ln -sf ~/.vim/tmux.conf ~/.tmux.conf
-git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
 
-#install vim plugin
-tmp_file=ivim_temp
-echo "ivim正在努力为您安装bundle程序" > $tmp_file
-echo "安装完毕将自动退出" >> $tmp_file
-echo "请耐心等待" >> $tmp_file
-vim $tmp_file -c "BundleInstall" -c "q" -c "q"
-rm $tmp_file
-echo "安装完成"
+#
+if [ ! -d ~/.vim/bundle ];then
+  git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+
+  #install vim plugin
+  tmp_file=ivim_temp
+  echo "ivim正在努力为您安装bundle程序" > $tmp_file
+  echo "安装完毕将自动退出" >> $tmp_file
+  echo "请耐心等待" >> $tmp_file
+  vim $tmp_file -c "BundleInstall" -c "q" -c "q"
+  rm $tmp_file
+  echo "安装完成"
+fi
